requirements.txt
allure-pytest==2.9.45
pytest~=7.1.2
selene>=2.0.b5
selenium~=4.2.0
python-dotenv~=0.20.0
